+++ src/property.ts
@@ -18,9 +18,9 @@ export class Property {
   type: Type;
   min: number | Date;
   max: number | Date;
-  send: boolean;
-  need: Method[];
-  typeCheck: boolean;
+  isPrivate: boolean;
+  requiredFor: Method[];
+  isTypeChecked: boolean;
   sanitizer: ((v:any) => any) | null;
   normalizer: ((v:any) => any) | null;
   validator: ((v:any) => any) | null;
@@ -30,9 +30,9 @@ export class Property {
     type: Type,
     min: number | Date | null,
     max: number | Date | null,
-    send: boolean,
-    need: Method[],
-    typeCheck: boolean,
+    isPrivate: boolean,
+    requiredFor: Method[],
+    isTypeChecked: boolean,
     sanitizer: ((v:any) => any) | null,
     normalizer: ((v:any) => any) | null,
     validator: ((v:any) => any) | null,
@@ -50,12 +50,12 @@ export class Property {
       throw new Error(`${LOGS_PREFIX}Property "type" must be a valid type - caused by: ${(err as Error).message}`);
     }
     
-    if (isArray(need)){
-      for (const m of need) {
+    if (isArray(requiredFor)) {
+      for (const m of requiredFor) {
         try {
           isIn(METHODS as unknown as unknown[], m, 0, true);
         } catch (err) {
-          throw new Error(`${LOGS_PREFIX}Property "need" must be an array of REST methods - caused by: ${(err as Error).message}`);
+          throw new Error(`${LOGS_PREFIX}Property "requiredFor" must be an array of REST methods - caused by: ${(err as Error).message}`);
         }
       }
     }
@@ -64,9 +64,9 @@ export class Property {
     this.type = type;
     this.min = this.interval(min, type, 0, "1900-01-01T00:00:00Z");
     this.max = this.interval(max, type, 999999999, "2200-12-31T00:00:00Z");
-    this.need = isArray(need) ? need : [];
-    this.send = isBoolean(send) ? send : true;
-    this.typeCheck = isBoolean(typeCheck) ? typeCheck : false;
+    this.requiredFor = isArray(requiredFor) ? requiredFor : [];
+    this.isPrivate = isBoolean(isPrivate) ? isPrivate : false;
+    this.isTypeChecked = isBoolean(isTypeChecked) ? isTypeChecked : false;
     this.sanitizer = isFunction(sanitizer) ? sanitizer : null;
     this.normalizer = isFunction(normalizer) ? normalizer : null;
     this.validator = isFunction(validator) ? validator : null;